Refactor ColorPAL sensor to individual module.
