+++ src/modules/editor.js
@@ -43,12 +43,15 @@ import './blockly/generators/propc/procedures';
 import './blockly/generators/propc/s3';
 import './blockly/generators/propc/sd_card';
 import './blockly/generators/propc/sensors';
+
+import './blockly/generators/propc/sensors/color_pal';
 import './blockly/generators/propc/sensors/dht22';
 import './blockly/generators/propc/sensors/mma7455';
 import './blockly/generators/propc/sensors/mx2125';
 import './blockly/generators/propc/sensors/ping';
 import './blockly/generators/propc/sensors/pir';
 import './blockly/generators/propc/sensors/sony_remote';
+
 import './blockly/generators/propc/variables';
 
 import {
